Move inactive workplaces endpoints into new file
